+++ app.js
@@ -5,29 +5,7 @@ App({
   onLaunch: function () {
     // tools.test();
     
-    wx.login({
-      success: function(res){
 
-        var data = {
-          "jsCode": res.code
-        };
-        api.userIfExist(data, function(res) {
-          console.log(res);
-          console.log(res.data.msg.openid);
-          api.openId = res.data.msg.openid;
-          //如果用户不存在就跳到guide页面
-          if(res.data.result=='exist')
-            wx.reLaunch({
-              url: '/pages/index/index'
-            });
-          //如果用户存在就跳到index页面
-          else if(res.data.result=='notExist')
-            wx.reLaunch({
-              url: '/pages/guide/guide'
-            });
-        });
-      }
-    })
     // 登录
     wx.login({
       success: res => {
