+++ pages/index/index.wxss
@@ -7,7 +7,7 @@
   height: 100%;
   margin:0rpx;
   padding:0rpx;
-  background: rgb(240, 237, 237);
+  background: rgb(248, 251, 252);
 }
 .container>scroll-view{
   width: 100%;
